Move SQLImplicits to sql/api
