+++ sql/core/src/test/scala/org/apache/spark/sql/test/SQLTestData.scala
@@ -34,9 +34,7 @@ private[sql] trait SQLTestData { self =>
   protected def spark: SparkSession
 
   // Helper object to import SQL implicits without a concrete SparkSession
-  private object internalImplicits extends SQLImplicits {
-    protected override def session: SparkSession = self.spark
-  }
+  private object internalImplicits extends SQLImplicits(self.spark)
 
   import internalImplicits._
   import SQLTestData._
